requirements.txt
dash==1.20.0
gunicorn==20.1.0
h5py==2.10.0
tensorflow-cpu
tflite
kapre==0.3.5
librosa==0.8.0
pandas #==1.2.4
plotly==4.14.3
requests==2.25.1
SoundFile==0.10.3.post1
requests
